int to long ended

+++ distribution.h
@@ -9,10 +9,10 @@ long **remainderR(long **secretS, long **projectionSd, int n);
 
 long **remainderRw(long **watermarkW, long **projectionSd, int n);
 
-int *g_i_j(int **R, int initial_column, int t, int n, int k);
+long *g_i_j(long **R, int initial_column, int t, int n, int k);
 
-int **matG_t(int **R, int n, int k, int t);
+long **matG_t(long **R, int n, int k, int t);
 
-int ***matG(int **R, int n, int k);
+long ***matG(long **R, int n, int k);
 
 #endif //CRIPTO_TP_DISTRIBUTION_H
