+++ modular.h
@@ -1,7 +1,7 @@
 #ifndef CRIPTO_TP_MODULAR_H
 #define CRIPTO_TP_MODULAR_H
 
-int modulo(int x, int N);
+long modulo(long x, int N);
 
 int *modularInverse(int n, int prime);
